Working on "focus" feature and more
Made it so users can click on a point on a line or the corresponding box for that city in the legend and "focus" that line/city by reducing the opacity of the other lines.

Also reformatted tooltip, fixed issue with y-axis and axis labels being re-drawn on top of themselves every time the chart updated, and added a select feature to allow users to look at the max or min daily temperature instead of just the mean (still need to make it functional though).